add AnonymousSubject
RxLua only has a zero-version create() function for Subject, corresponding to
the constructor of RxJS for example.  However, RxJS has a two-argument
create() function, returning a subject that simply juxtaposes an observable
and an observer into a single object.  Quite confusingly, the object returned
by create() is not a Subject, but an instance of a different class
AnonymousSubject.  Introduce AnonymousSubject as a completely separate
implementation to avoid overloading Subject.create.